Reorganize autodiscover files
