+++ pkg/autodiscover/autodiscover_nads_test.go
@@ -1,0 +1,1 @@
+package autodiscover
